dimControl: Tools for Geometric Processing and Dimensional Analysis

A collection of tools for geometric processing, reconstruction, and dimensional analysis of three-dimensional data.

Main functionality

dimControl provides functionality for point cloud and mesh processing, boundary extraction, angle estimation, component detection, surface reconstruction, and geometric analysis.

Getting started

Work in progress

See also

Author

Maintainer: Nataly Romaris nataly.romaris@udc.es

Authors:

Olamar Benavente

Ruben Fernandez-Casal

Salvador Naya

Javier Tarrio-Saavedra