Simulate floor point cloud

Simulates a floor point cloud around the bounding box of a 3D object point cloud.

Usage

simulateFloor(object, n, gap, margin, rgen = .rtnormDefault, ...)

Arguments

object

Numeric matrix with three columns (X, Y, Z) containing the object point cloud.

n

Number of floor points to simulate.

gap

Non-negative numeric value specifying the vertical distance between the lowest point of the object and the reference floor height.

margin

Positive numeric value specifying the width of the floor region around the object in the X and Y directions.

rgen

Random number generator used to perturb the floor height. By default, a truncated normal distribution is used.

...

Additional arguments passed to rgen.

Value

A numeric matrix with n rows and three columns (X, Y, Z) containing the simulated floor point cloud.

Details

The floor is simulated in four rectangular regions surrounding the X-Y bounding box of object: left, right, bottom, and top.

The reference floor height is defined as the minimum Z coordinate of the object minus gap. Random noise generated by rgen is then added to this reference height.

By default, the noise follows a truncated normal distribution with mean 0, standard deviation 0.2, and lower and upper bounds of -0.5 and 0.5, respectively.